convert utils to ts
