+++ README.md
@@ -146,5 +146,4 @@ When a pull request is merged into the `main` branch, the following is automatic
 
 This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-
 ---
