+++ package.json
@@ -34,6 +34,7 @@
     "@types/jest-axe": "^3.5.3",
     "@types/redux-mock-store": "^1.0.3",
     "@types/styled-components": "5.1.15",
+    "@types/node": "16.11.7",
     "@typescript-eslint/eslint-plugin": "^5.3.1",
     "@typescript-eslint/parser": "^5.3.1",
     "axios-mock-adapter": "^1.20.0",
@@ -54,7 +55,7 @@
     "react-scripts": "^4.0.3",
     "redux-mock-store": "^1.5.4",
     "sass": "1.39.2",
-    "stylelint": "^14.0.1",
+    "stylelint": "^13.12.0",
     "stylelint-config-prettier": "^9.0.3",
     "typescript": "^4.4.4",
     "uglify-js": "^3.14.3"
@@ -90,9 +91,9 @@
     "lint:js": "tsc --noEmit && esw --ext .js,.jsx,.ts,.tsx --color ",
     "lint:js:fix": "npm run lint:js -- --fix",
     "prettier": "prettier  --ignore-path ./.eslintignore --check",
-    "prettier:all": "npm run prettier '**/*.+(js|jsx|ts|tsx|html|css|scss|json)'",
+    "prettier:all": "npm run prettier '**/*.+(js|jsx|ts|tsx|css|scss|html|json)'",
     "prettier:fix": "prettier  --ignore-path ./.eslintignore --write",
-    "prettier:fix:all": "npm run prettier:fix '**/*.+(js|jsx|ts|tsx|html|css|scss|json)'",
+    "prettier:fix:all": "npm run prettier:fix '**/*.+(js|jsx|ts|tsx|css|scss|html|json)'",
     "lint:scss": "stylelint --ignore-path ./.eslintignore ",
     "lint:scss:all": "npm run lint:scss './**/*.+(css|scss)'",
     "lint:scss:fix:all": "npm run lint:scss -- --fix './**/*.+(css|scss)'",
