+++ utils/resetPackageJson.ts
@@ -1,5 +1,8 @@
 // This file minimizes the package.json file before it is included in the package to be published.
+
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 /* eslint-disable @typescript-eslint/no-var-requires */
+// @ts-ignore
 const path = require('path');
 const fs = require('fs');
